[pdf] Expand content item scope
In a tagged PDF drawing operations should be associated with the
document structure tree. This is done by drawing on a page inside a
marked-content sequence with a marked-content identifier which is used
as a content item in a structure element. Previously a new
marked-content sequence with a new marked-content identifier was created
for each image and run of glyphs drawn. This is often excessive, as it
is common for multiple text draws in a row to be part of a single
structure element. In addition, the existing code did not associate path
draws with the currently active structure element.

Instead, open and close the marked-content sequences only as needed.
Open a marked-content sequence when there is a current structure
element and there is no current marked-content sequence or the current
marked-content sequence isn't associated with the correct structure
element. Close a marked-content sequence when openening a new
marked-content sequence, when about to draw and there is no current
structure element, or at the end of the page.

Change-Id: Idc08821ecc843f416331eb8fb703d578ffb77949
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/909876
Commit-Queue: Ben Wagner <bungeman@google.com>
Reviewed-by: Florin Malita <fmalita@google.com>